--- InsuranceAdminService/src/main/java/com/insurance/adminservice/exception/ServiceCenterIdIsNotFoundException.java
@@ -1,12 +1,0 @@
-package com.insurance.adminservice.exception;
-
-public class ServiceCenterIdIsNotFoundException extends RuntimeException {
-
-	
-	public ServiceCenterIdIsNotFoundException(String msg)
-	{
-		super(msg);
-	}
-	
-	
-}
