+++ InsuranceAdminService/src/main/java/com/insurance/adminservice/model/Employee.java
@@ -22,7 +22,6 @@ public class Employee {
 	private long employeeContactNumber;
 	private float employeeSalary;
 	private String employeeDesignation;
-	private String employeeEmailId;
 	private String username;
 	private String password;
