+++ InsuranceAdminService/src/main/java/com/insurance/adminservice/servicei/AdminServiceI.java
@@ -2,12 +2,8 @@ package com.insurance.adminservice.servicei;
 
 import com.insurance.adminservice.model.Employee;
 
-public interface EmployeeServiceI {
+public interface AdminServiceI {
 
 	Employee saveEmployee(Employee employee);
-
-	void removeEmployeeById(int empoyeeId);
-
-	Employee updateEmployeeDataById(Employee emp, int employeeId);
 
 }
