+++ InsuranceAdminService/src/main/java/com/insurance/adminservice/servicei/EmployeeService.java
@@ -1,0 +1,5 @@
+package com.insurance.adminservice.servicei;
+
+public interface EmployeeService {
+
+}
