+++ InsuranceAdminService/src/main/java/com/insurance/adminservice/servicei/ServiceCenterService.java
@@ -6,8 +6,4 @@ public interface ServiceCenterService {
 
 	ServiceCenter saveServiceCenter(ServiceCenter serviceCenter);
 
-	void removeServiceCenterById(int serviceCenterId);
-
-	ServiceCenter UpdateServiceCenterDataById(ServiceCenter service, int servicecenterId);
-
 }
