+++ InsuranceAdminService/src/main/java/com/insurance/adminservice/serviceimpl/ServiceCenterImpl.java
@@ -1,11 +1,8 @@
 package com.insurance.adminservice.serviceimpl;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.insurance.adminservice.exception.ServiceCenterIdIsNotFoundException;
 import com.insurance.adminservice.model.ServiceCenter;
 import com.insurance.adminservice.repository.ServiceCenterRepository;
 import com.insurance.adminservice.servicei.ServiceCenterService;
@@ -19,42 +16,5 @@ public class ServiceCenterImpl implements ServiceCenterService{
 	public ServiceCenter saveServiceCenter(ServiceCenter serviceCenter) {
 		
 		return repository.save(serviceCenter);
-	}
-
-	@Override
-	public void removeServiceCenterById(int serviceCenterId) {
-		
-	Optional<ServiceCenter>service=repository.findById(serviceCenterId);
-		if(service.isPresent())
-		{
-			repository.deleteById(serviceCenterId);
-		}
-		
-		else
-		{
-			
-			throw new ServiceCenterIdIsNotFoundException("Service Center Id   "+serviceCenterId+"  Is Not Present");
-		}
-		
-		
-	}
-
-	@Override
-	public ServiceCenter UpdateServiceCenterDataById(ServiceCenter service, int servicecenterId) {
-
-		Optional<ServiceCenter>servicecenter=repository.findById(servicecenterId);
-		if(servicecenter.isPresent())
-		{
-			
-			return repository.save(service);
-			
-		}
-		else
-		{
-			
-			throw new ServiceCenterIdIsNotFoundException("ServiceCenter Id    "+servicecenterId+"    Is NOt Present");
-		}
-		
-		
 	}
 }
